fix(ci): align executor versions with update metadata

See merge request weibo_rd/common/wecode/wegent!3221

+++ wecode/docker/executor/prepare_build.sh
@@ -2,8 +2,11 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 
+: "${EXECUTOR_VERSION:?missing resolved Executor version}"
+
 rm -rf ./executor ./sdk ./shared
 cp -r ../../../executor ./executor
+printf '%s\n' "$EXECUTOR_VERSION" > ./executor/.build-version
 mkdir -p ./sdk
 cp -r ../../../sdk/plugin-auth ./sdk/plugin-auth
 cp -r ../../../sdk/plugin-creator ./sdk/plugin-creator
